feat: UIViewController에 공통 Alert 기능 확장 추가 및 네트워크 에러 처리 추가
